+++ pandas/core/arrays/sparse/__init__.py
@@ -1,0 +1,11 @@
+# flake8: noqa: F401
+
+from .sparse import (
+    BlockIndex,
+    IntIndex,
+    SparseAccessor,
+    SparseArray,
+    SparseDtype,
+    SparseFrameAccessor,
+    _make_index,
+)
